requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
streamlit
google-generativeai
